api/requirements.txt
# Django 和 DRF 核心依赖
Django==5.0.2
djangorestframework==3.14.0

# CORS 支持
django-cors-headers==4.3.1

# 认证相关
djangorestframework-simplejwt==5.3.1

# 数据库驱动和连接池
mysqlclient==2.2.4      # MySQL
django-mysql==4.12.0    # MySQL 增强功能
django-db-connection-pool==1.2.4  # 数据库连接池

# 缓存支持
django-redis==5.4.0     # Redis 缓存后端
redis==5.0.1           # Redis 客户端

# 文件上传和图片处理
Pillow==10.2.0
python-magic==0.4.27      # 文件类型检测

# Excel处理
openpyxl==3.1.2         # Excel文件处理

# API 文档
drf-spectacular==0.28.0

# 开发工具
django-debug-toolbar==4.3.0

# 时间处理
pytz==2024.1

# 环境变量管理
python-dotenv==1.0.1

# 密码哈希
bcrypt==4.1.2

# 工具库
requests==2.31.0

# 性能监控
django-prometheus==2.3.1  # Prometheus 监控集成 

# WSGI server
gunicorn==21.2.0 